update create config script

+++ wis2box-create-config.py
@@ -28,6 +28,27 @@ import string
 from string import Template
 from typing import Tuple
 
+def get_country_name(country_code: str) -> str:
+    """
+    provide the country name for the wis2box
+    using the country's 3-letter ISO code
+    using the data from config-templates/bounding_box_lookup.json
+    """
+    
+    country_name = 'NA'
+    # get the path to the data
+    data_path = Path(__file__).parent / 'config-templates' / 'countries.json'
+    # open the file
+    with data_path.open() as fh:
+        # load the data
+        data = json.load(fh)
+        # get the bounding box for the country
+        if country_code in data['countries'] and 'name' in data['countries'][country_code]:  # noqa
+            country_name = data['countries'][country_code]['name']
+        else:
+            print(f'No country name found for "{country_code}".')
+            print('Using "NA" as the country name.')
+    return country_name
 
 def get_bounding_box(country_code: str) -> Tuple[str, str]:
     """
@@ -96,7 +117,7 @@ def get_bounding_box(country_code: str) -> Tuple[str, str]:
 
 def get_country_and_centre_id() -> Tuple[str, str]:
     """
-    Asks the user for the 3-letter ISO country-code
+    Asks the user for the 2-letter ISO country-code
     and a string identifying the centre hosting the wis2box.
 
     :returns: `tuple` of (country_code, centre_id)
@@ -108,33 +129,48 @@ def get_country_and_centre_id() -> Tuple[str, str]:
         if answer == 'exit':
             exit()
 
-        print('Please enter your 3-letter ISO country code:')
+        print('Please enter your 2-letter ISO country code:')
         country_code = input()
+        country_name = 'NA'
 
         # check that the input is a 3-letter string
         # if not repeat the question
-        while len(country_code) != 3:
-            print('The country code must be a 3-letter string.')
-            print('Please enter your 3-letter ISO country code:')
+        while len(country_code) != 2:
+            print('The country code must be a 2-letter string.')
+            print('Please enter your 2-letter ISO country code:')
             country_code = input()
+            country_name = get_country_name(country_code)
+            answer = ""
+            while answer != "y":
+                if answer == "exit":
+                    exit()
+                print("The country name will be set to:")
+                print(f"    {country_name}")
+                print("Is this correct? (y/n/exit)")
+                answer = input()
 
         # make sure the country code is lowercase
         country_code = country_code.lower()
 
         print('Please enter the centre-id for your wis2box:')
         centre_id = str(input()).lower()
+        # check that centre_id starts with the country_code followed by a dash # noqa
+        # if not repeat the question
+        while not centre_id.startswith(country_code + '-'):
+            print('The centre-id must start with the country code followed by a dash.')
+            print(f"For example: {country_code}-{centre_id}")
+            print('Please enter the string identifying the centre hosting the wis2box:') # noqa
+            centre_id = str(input()).lower()
 
         # check that the input is valid
         # if not repeat the question
-        while any([x in centre_id for x in ['#', '+', ' ']]) or len(centre_id) < 3:  # noqa
-            print('The centre-id cannot contain spaces or the "+" or "#" characters, and must be at least 3 characters long.') # noqa
+        while any([x in centre_id for x in ['#', '+', ' ']]) or len(centre_id) < 5:  # noqa
+            print('The centre-id cannot contain spaces or the "+" or "#" characters, and must be at least 5 characters long.') # noqa
             print('Please enter the string identifying the centre hosting the wis2box:') # noqa
             centre_id = str(input()).lower()
 
         # ask the user to confirm their choice and give them the option to change it # noqa
         # and give them the option to exit the script
-        print('The country-code will be set to:')
-        print(f'  {country_code}')
         print('The centre-id will be set to:')
         print(f'  {centre_id}')
         print('Is this correct? (y/n/exit)')
@@ -368,7 +404,7 @@ def create_config_dir() -> str:
     return config_dir
 
 
-def create_datamappings_file(config_dir: str, country_code: str,
+def update_datamappings_file(config_dir: str, country_code: str,
                              centre_id: str) -> None:
     """
     creates the data mappings file in the directory config_dir
@@ -391,18 +427,25 @@ def create_datamappings_file(config_dir: str, country_code: str,
     with template_file.open() as fh:
         config_file = Template(fh.read())
         result = config_file.substitute(template_vars)
-        with new_config_file.open("w") as fh2:
-            fh2.write(result)
+        if new_config_file.exists():
+            print(f"Adding datasets to {new_config_file}")
+            with new_config_file.open("a") as fh2:
+                fh2.write(result.replace("data:\n",""))
+        else:
+            print(f"Creating {new_config_file}")
+            with new_config_file.open("w") as fh2:
+                fh2.write(result)
 
     print("*" * 80)
-    print("Initial data_mappings.yml have been created") # noqa
+    print("data_mappings.yml have been created/updated") # noqa
     print("Please review the file and update as needed.")
     print("*" * 80)
 
 
 def create_metadata_file(config_dir: str, country_code: str, country_name,
                          centre_id: str, centre_name: str, wis2box_email: str,
-                         bounding_box: str, template: str) -> str:
+                         contact_name: str, bounding_box: str,
+                         template: str) -> str:
     """
     creates the metadata file in the directory config_dir
 
@@ -412,6 +455,7 @@ def create_metadata_file(config_dir: str, country_code: str, country_name,
     :param centre_id: `str` of the centre id of the wis2box
     :param centre_name: `str` of centre name of the wis2box
     :param wis2box_email: `str` of centre email
+    :param contact_name: `str` of contact name
     :param bounding_box: `str` of CSV of bounding box
     :param template: `str` of synop or temp
     
@@ -428,7 +472,7 @@ def create_metadata_file(config_dir: str, country_code: str, country_name,
     if not discovery_metadata_dir.exists():
         discovery_metadata_dir.mkdir(parents=True)
 
-    new_config_file = discovery_metadata_dir / f'metadata-{template}.yml'
+    new_config_file = discovery_metadata_dir / f'metadata-{template}-{centre_id}.yml'
     template_file = Path('config-templates') / f'metadata-{template}.yml.tmpl'  # noqa
 
     template_vars = {
@@ -440,7 +484,8 @@ def create_metadata_file(config_dir: str, country_code: str, country_name,
         'CENTRE_ID': centre_id,
         'CENTRE_NAME': centre_name,
         'WIS2BOX_EMAIL': wis2box_email,
-        'BOUNDING_BOX': bounding_box
+        'BOUNDING_BOX': bounding_box,
+        'CONTACT_NAME': contact_name
     }
 
     with template_file.open() as fh:
@@ -473,23 +518,35 @@ def create_metadata_files(config_dir: str, country_code: str,
     while answer != "y":
         if answer == "exit":
             exit()
-
-        print("Please enter the email address of the wis2box administrator:")
+        print(f"Please enter the email address of the wis2box administrator for {centre_id}:")
         wis2box_email = input()
         print("The email address of the wis2box administrator will be set to:") # noqa
         print(f"    {wis2box_email}")
         print("Is this correct? (y/n/exit)")
         answer = input()
 
-    # ask for the name of the centre
+    # ask for the contact name of the wis2box administrator
     answer = ""
-
+    wis2box_contact_name = ""
+    
     while answer != "y":
         if answer == "exit":
             exit()
-        print("Please enter the name of your organization:")
+        print(f"Please enter the name of the wis2box administrator for {centre_id}:") # noqa
+        wis2box_contact_name = input()
+        print("The name of the wis2box administrator will be set to:")
+        print(f"    {wis2box_contact_name}")
+        print("Is this correct? (y/n/exit)")
+        answer = input()
+
+    # ask for the name of the centre
+    answer = ""
+    while answer != "y":
+        if answer == "exit":
+            exit()
+        print(f"Please enter the name of the organization for {centre_id}:")
         centre_name = input()
-        print("Your organization name will be set to:")
+        print("The organization name will be set to:")
         print(f"    {centre_name}")
         print("Is this correct? (y/n/exit)")
         answer = input()
@@ -497,29 +554,22 @@ def create_metadata_files(config_dir: str, country_code: str,
     # get an initial bounding box for the country
     country_name, bounding_box = get_bounding_box(country_code)
 
-    create_metadata_file(
-        config_dir,
-        country_code,
-        country_name,
-        centre_id,
-        centre_name,
-        wis2box_email,
-        bounding_box,
-        template="synop"
-    )
-    create_metadata_file(
-        config_dir,
-        country_code,
-        country_name,
-        centre_id,
-        centre_name,
-        wis2box_email,
-        bounding_box,
-        template="temp"
-    )
+    templates = ['synop', 'temp']
+    for template in templates:
+        create_metadata_file(
+            config_dir,
+            country_code,
+            country_name,
+            centre_id,
+            centre_name,
+            wis2box_email,
+            wis2box_contact_name,
+            bounding_box,
+            template=template
+        )
 
     print("*" * 80)
-    print(f"Initial metadata files created in directory {config_dir}.") # noqa
+    print(f"Metadata files for {centre_id} created in directory {config_dir}.") # noqa
     print("Please review the files and edit where necessary.")
     print("*" * 80)
 
@@ -616,14 +666,21 @@ def main():
     if not dev_env.is_file():
         create_wis2box_env(config_dir)
 
-    country_code, centre_id = get_country_and_centre_id()
-
+    q_metadata = 'Would you like to add datasets for another country/centre_id ? (y/n)' 
     print("*" * 80)
-    print("Creating initial configuration for surface and upper-air data.")
+    print("Preparing datasets for your country/centre_id.")
     print("*" * 80)
+    answer = 'y'
+    while answer == 'y':
+        country_code, centre_id = get_country_and_centre_id()
+        create_metadata_files(config_dir, country_code, centre_id)
+        update_datamappings_file(config_dir, country_code, centre_id)
+        print(q_metadata)
+        answer = input()
+        if answer not in ['y', 'n']:
+            print(q_metadata)
+            answer = input()
 
-    create_metadata_files(config_dir, country_code, centre_id)
-    create_datamappings_file(config_dir, country_code, centre_id)
     create_station_list(config_dir)
 
     print("The configuration is complete.")
